+++ shepherd_core/shepherd_core/data_models/task/emulation.py
@@ -67,7 +67,6 @@ class EmulationTask(ShpModel):
     """ timestamp or unix epoch time, None = ASAP"""
     duration: Optional[timedelta] = None
     """ ⤷ Duration of recording in seconds, None = till EOF"""
-    abort_on_error: Annotated[bool, deprecated("has no effect")] = False
 
     # emulation-specific
     use_cal_default: bool = False