+++ shepherd_core/shepherd_core/data_models/task/harvest.py
@@ -40,7 +40,6 @@ class HarvestTask(ShpModel):
     """ timestamp or unix epoch time, None = ASAP"""
     duration: Optional[timedelta] = None
     """ ⤷ Duration of recording in seconds, None = till EOFSys"""
-    abort_on_error: Annotated[bool, deprecated("has no effect")] = False
 
     # emulation-specific
     use_cal_default: bool = False