+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v2025.06.2
 
-- remove deprecated fields in Experiment, Systemlogging, Tasks, ObserverTasks, TestbedTasks
+- remove deprecated fields in Experiment, Systemlogging, Tasks, ObserverTasks, TestbedTasks, EmulationTask, HarvestTask
+- remove deprecated Reader.buffers_n 
 - allow configuring battery in vsource
 - allow configuring energy environment with multiple recordings